+++ lib/src/views/otp_view/state/otp_cubit.dart
@@ -11,14 +11,14 @@ class OtpCubit extends Cubit<OtpState> {
 
   UserModel? userModel;
 
-  Future<void> requestOtp(String phoneNumber) async {
+  Future<void> requestOtp({required String phoneNumber,required String countyCode, required String e164Key }) async {
     // setting up null if user tries to login again
     // then the sucess might get calls even if current request has failed.
     userModel = null;
     emit(OtpLoading(DateTime.now()));
 
     try {
-      userModel = await service.phoneAuth(phoneNumber);
+      userModel = await service.phoneAuth(phoneNumber: phoneNumber, countyCode: countyCode, e164Key: e164Key);
       if (userModel == null) {
         emit(OtpError('Some error occured.', DateTime.now()));
       } else {